Split builtin.rs into category-based modules
Decompose the monolithic builtin.rs into:
- builtin/mod.rs: thin dispatcher (0.34)
- builtin_math.rs: math functions (3.82)
- builtin_list.rs: list/statistics functions (21.71)
- builtin_constructor.rs: type constructors (7.97)
- builtin_time.rs: time utilities (5.19)
- builtin_helpers.rs: shared helpers (24.43)

All files now under 25 complexity, down from 66.07.
